Refactor : make uils file AddRecentProduct, HandleProduct component 2번 이상 사용되는 함수는 utils 폴더로 빼서 사용한다.
