+++ src/Pages/ProductList/utils/AddRecentProduct.js
@@ -1,0 +1,5 @@
+export function AddRecentProduct(recentItems, ClickProd) {
+  recentItems.push(ClickProd);
+  const stringProds = JSON.stringify(recentItems);
+  localStorage.setItem("recentItems", stringProds);
+}
